# swagger.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: ViMbAdmin-API
  description: This is a quick and dirty Lumen base API for the ViMbAdmin database
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
host: vimbadmin-api.dev
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  oAuthSecurity:
    type: oauth2
    tokenUrl: https://vimbadmin-api.dev/oauth/token
    flow: application
    scopes:
      default: Default
paths:
...
  /domains/{id}:
    get:
      description: Return single domains
      operationId: getDomain
      produces:
        - application/json
      security:
        - oAuthSecurity:
          - default
      parameters:
        - $ref: '#/parameters/id'
        - $ref: '#/parameters/include'
      responses:
        '200':
          description: domain response
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                 $ref: '#/definitions/domain'
              included:
                $ref: "#/definitions/included"
              links:
                 $ref: '#/definitions/link'
        '401':
          $ref: '#/responses/401'
        '404':
          $ref: '#/responses/404'
        default:
          $ref: '#/responses/default'
...
parameters:
...
  id:
    name: id
    in: path
    description: id
    required: true
    type: string
...
  include:
    name: include
    in: query
    description: additional relations to include in the response
    required: false
    type: array
    items:
      enum: ['mailboxes', 'aliases']
    uniqueItems: true
    collectionFormat: csv
responses:
  401:
    description: Unauthorized
    schema:
      type: object
      required:
        - errors
      properties:
        errors:
          $ref: '#/definitions/error'
        links:
          $ref: '#/definitions/link'
  404:
    description: Not Found
    schema:
      type: object
      required:
        - errors
      properties:
        errors:
          $ref: '#/definitions/error'
        links:
          $ref: '#/definitions/link'
...
  default:
    description: unexpected error
    schema:
      type: object
      required:
        - errors
      properties:
        errors:
          $ref: '#/definitions/error'
        links:
          $ref: '#/definitions/link'
definitions:
  domain:
    type: object
    required:
      - id
      - type
    properties:
      id:
        type: integer
        format: int64
      type:
        type: string
      attributes:
        type: object
        properties:
          domain:
            type: string
      links:
        $ref: '#/definitions/link'
      relationships:
        $ref: "#/definitions/relationships"
...
  relationships:
    type: object
    properties:
      domains:
        $ref: "#/definitions/relation"
      aliases:
        $ref: "#/definitions/relation"
      mailboxes:
        $ref: "#/definitions/relation"
  relation:
    type: object
    properties:
      links:
        $ref: "#/definitions/link"
      data:
        type: array
        items:
          type: object
          required:
            - id
            - type
          properties:
            id:
              type: integer
              format: int64
            type:
              type: string
  included:
    type: array
    items:
      type: object
      required:
        - id
        - type
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
        attributes:
          type: object
        links:
          $ref: '#/definitions/link'
  link:
    type: object
    properties:
      self:
        type: string
      related:
        type: string
  error:
    type: object
    required:
      - status
      - title
      - detail
    properties:
      status:
        type: integer
      title:
        type: string
      detail:
        type: string
      meta:
        type: object
...